+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [0.68.0]
 
+* Add chat participant `/tests` slash command for getting help with writing and running tests
 * Add chat participant `/workspace` slash command for searching workspace-specific documentation
 * Update the chat participant to also include recent chat history in requests to the language model
 * Explicitly dispose of the Logtalk terminal to prevent restoration when the extension is deactivated